finish renaming and updating messages to indicate we are supporting both GAR and GCR with this provider
